Refactor: change place of files quize
